Unit tests for components services migration (#2762)
* Unit tests

* fix e2e failures

* Final batch of components services migration (#2763)

* storageManagerFactory

* playlistComponentFactory

* unit tests
[stage-1]
