+++ src/app/ajs-upgraded-providers.spec.ts
@@ -15,7 +15,7 @@ import {
   PresentationUtils, presentationUtilsProvider,
   UserAuthFactory, userAuthFactoryProvider,
   CompanySettingsFactory, companySettingsFactoryProvider,
-  UserState, userStateProvider, purchaseFlowTrackerProvider, PurchaseFlowTracker, contactServiceProvider, ContactService, PresentationService, presentationServiceProvider, createFirstScheduleServiceProvider, CreateFirstScheduleService, scheduleFactoryProvider, ScheduleFactory, PresentationTracker, presentationTrackerProvider, ScheduleSelectorFactory, scheduleSelectorFactoryProvider, StorageUtils, storageUtilsProvider, StorageAPILoader, storageAPILoaderProvider, UpdateCompany, updateCompanyProvider
+  UserState, userStateProvider, purchaseFlowTrackerProvider, PurchaseFlowTracker, contactServiceProvider, ContactService, PresentationService, presentationServiceProvider, createFirstScheduleServiceProvider, CreateFirstScheduleService, scheduleFactoryProvider, ScheduleFactory, PresentationTracker, presentationTrackerProvider, ScheduleSelectorFactory, scheduleSelectorFactoryProvider, StorageUtils, storageUtilsProvider, StorageAPILoader, storageAPILoaderProvider, UpdateCompany, updateCompanyProvider, StorageService, storageServiceProvider, ScrollingListService, scrollingListServiceProvider
 } from './ajs-upgraded-providers';
 
 describe('ajs-upgraded-providers', () => {
@@ -157,6 +157,16 @@ describe('ajs-upgraded-providers', () => {
   it('updateCompany:', () => {
     testRegisterProvider(updateCompanyProvider, UpdateCompany);
     testAngularJsService(updateCompanyProvider, 'updateCompany');
+  });
+
+  it('storage:', () => {
+    testRegisterProvider(storageServiceProvider, StorageService);
+    testAngularJsService(storageServiceProvider, 'storage');
+  });
+
+  it('ScrollingListService:', () => {
+    testRegisterProvider(scrollingListServiceProvider, ScrollingListService);
+    testAngularJsService(scrollingListServiceProvider, 'ScrollingListService');
   });
   
 });
